Create database connection
func connect() was created by using basic sql/package
for centralize code with db connection
this function return *sql.DB so that it can be used by func Gorm()
that implemented gorm/package, but benefit from using this approach in
way reuse main connection to database.

Example with Gorm:
	connection, err := connect()
	if err != nil {
		panic(err)
	}

	db, err := gorm.Open(mysql.New(mysql.Config{
		Conn: connection,
	}))

Example with SqlX:
	connection, err := connect()
	if err != nil {
		panic(err)
	}

	db, err := sqlx.NewDB(connection(), "mysql")

This code give an opportunity to change credential only in one file
and if change driver it won't make many problems.